requirements.txt
langchain_openai
python-dotenv
langchain_astradb
langchain_community
unstructured

fastapi
uvicorn